# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_bookmark-project-learn_user-service_internal_models_dto_api_user.UserLogin:
    properties:
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - username
    type: object
...
  internal_handler_user.loginResponse:
    properties:
      data:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a book API
  title: Book API
  version: "1.0"
paths:
...
  /v1/users/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: Input required
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_bookmark-project-learn_user-service_internal_models_dto_api_user.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_user.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler_user.loginResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler_user.loginResponse'
      summary: Login
      tags:
      - user
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"